Goodbye inputcheck
