+++ publicapi/collection.go
@@ -12,7 +12,6 @@ import (
 	"github.com/go-playground/validator/v10"
 	db "github.com/mikeydub/go-gallery/db/gen/coredb"
 	"github.com/mikeydub/go-gallery/graphql/dataloader"
-	"github.com/mikeydub/go-gallery/publicapi/inputcheck"
 	"github.com/mikeydub/go-gallery/service/persist"
 )
 
@@ -32,7 +31,7 @@ type CollectionAPI struct {
 
 func (api CollectionAPI) GetCollectionById(ctx context.Context, collectionID persist.DBID) (*db.Collection, error) {
 	// Validate
-	if err := inputcheck.ValidateFields(api.validator, inputcheck.ValidationMap{
+	if err := validate.ValidateFields(api.validator, validate.ValidationMap{
 		"collectionID": {collectionID, "required"},
 	}); err != nil {
 		return nil, err
@@ -49,7 +48,7 @@ func (api CollectionAPI) GetCollectionById(ctx context.Context, collectionID per
 func (api CollectionAPI) GetCollectionsByIds(ctx context.Context, collectionIDs []persist.DBID) ([]*db.Collection, []error) {
 	collectionThunk := func(collectionID persist.DBID) func() (db.Collection, error) {
 		// Validate
-		if err := inputcheck.ValidateFields(api.validator, inputcheck.ValidationMap{
+		if err := validate.ValidateFields(api.validator, validate.ValidationMap{
 			"collectionID": {collectionID, "required"},
 		}); err != nil {
 			return func() (db.Collection, error) { return db.Collection{}, err }
@@ -86,7 +85,7 @@ func (api CollectionAPI) GetCollectionsByIds(ctx context.Context, collectionIDs 
 
 func (api CollectionAPI) GetCollectionsByGalleryId(ctx context.Context, galleryID persist.DBID) ([]db.Collection, error) {
 	// Validate
-	if err := inputcheck.ValidateFields(api.validator, inputcheck.ValidationMap{
+	if err := validate.ValidateFields(api.validator, validate.ValidationMap{
 		"galleryID": {galleryID, "required"},
 	}); err != nil {
 		return nil, err
@@ -101,7 +100,7 @@ func (api CollectionAPI) GetCollectionsByGalleryId(ctx context.Context, galleryI
 }
 
 func (api CollectionAPI) CreateCollection(ctx context.Context, galleryID persist.DBID, name string, collectorsNote string, tokens []persist.DBID, layout persist.TokenLayout, tokenSettings map[persist.DBID]persist.CollectionTokenSettings, caption *string) (*db.Collection, *db.FeedEvent, error) {
-	fieldsToValidate := inputcheck.ValidationMap{
+	fieldsToValidate := validate.ValidationMap{
 		"galleryID":      {galleryID, "required"},
 		"name":           {name, "collection_name"},
 		"collectorsNote": {collectorsNote, "collection_note"},
@@ -113,13 +112,13 @@ func (api CollectionAPI) CreateCollection(ctx context.Context, galleryID persist
 	var trimmedCaption string
 	if caption != nil {
 		trimmedCaption = strings.TrimSpace(*caption)
-		fieldsToValidate["caption"] = inputcheck.ValWithTags{trimmedCaption, fmt.Sprintf("required,caption")}
+		fieldsToValidate["caption"] = validate.ValWithTags{trimmedCaption, fmt.Sprintf("required,caption")}
 		cleaned := validate.SanitizationPolicy.Sanitize(trimmedCaption)
 		caption = &cleaned
 	}
 
 	// Validate
-	if err := inputcheck.ValidateFields(api.validator, fieldsToValidate); err != nil {
+	if err := validate.ValidateFields(api.validator, fieldsToValidate); err != nil {
 		return nil, nil, err
 	}
 
@@ -193,7 +192,7 @@ func (api CollectionAPI) CreateCollection(ctx context.Context, galleryID persist
 
 func (api CollectionAPI) DeleteCollection(ctx context.Context, collectionID persist.DBID) error {
 	// Validate
-	if err := inputcheck.ValidateFields(api.validator, inputcheck.ValidationMap{
+	if err := validate.ValidateFields(api.validator, validate.ValidationMap{
 		"collectionID": {collectionID, "required"},
 	}); err != nil {
 		return err
@@ -214,7 +213,7 @@ func (api CollectionAPI) DeleteCollection(ctx context.Context, collectionID pers
 
 func (api CollectionAPI) UpdateCollectionInfo(ctx context.Context, collectionID persist.DBID, name string, collectorsNote string) error {
 	// Validate
-	if err := inputcheck.ValidateFields(api.validator, inputcheck.ValidationMap{
+	if err := validate.ValidateFields(api.validator, validate.ValidationMap{
 		"collectionID":   {collectionID, "required"},
 		"name":           {name, "collection_name"},
 		"collectorsNote": {collectorsNote, "collection_note"},
@@ -255,7 +254,7 @@ func (api CollectionAPI) UpdateCollectionInfo(ctx context.Context, collectionID 
 }
 
 func (api CollectionAPI) UpdateCollectionTokens(ctx context.Context, collectionID persist.DBID, tokens []persist.DBID, layout persist.TokenLayout, tokenSettings map[persist.DBID]persist.CollectionTokenSettings, caption *string) (*db.FeedEvent, error) {
-	fieldsToValidate := inputcheck.ValidationMap{
+	fieldsToValidate := validate.ValidationMap{
 		"collectionID": {collectionID, "required"},
 		"tokens":       {tokens, fmt.Sprintf("required,unique,min=1,max=%d", maxTokensPerCollection)},
 		"sections":     {layout.Sections, fmt.Sprintf("unique,sorted_asc,lte=%d,min=1,max=%d,len=%d,dive,gte=0,lte=%d", len(tokens), maxSectionsPerCollection, len(layout.SectionLayout), len(tokens)-1)},
@@ -265,13 +264,13 @@ func (api CollectionAPI) UpdateCollectionTokens(ctx context.Context, collectionI
 	var trimmedCaption string
 	if caption != nil {
 		trimmedCaption = strings.TrimSpace(*caption)
-		fieldsToValidate["caption"] = inputcheck.ValWithTags{trimmedCaption, fmt.Sprintf("required,caption")}
+		fieldsToValidate["caption"] = validate.ValWithTags{trimmedCaption, fmt.Sprintf("required,caption")}
 		cleaned := validate.SanitizationPolicy.Sanitize(trimmedCaption)
 		caption = &cleaned
 	}
 
 	// Validate
-	if err := inputcheck.ValidateFields(api.validator, fieldsToValidate); err != nil {
+	if err := validate.ValidateFields(api.validator, fieldsToValidate); err != nil {
 		return nil, err
 	}
 
@@ -323,7 +322,7 @@ func (api CollectionAPI) UpdateCollectionTokens(ctx context.Context, collectionI
 
 func (api CollectionAPI) UpdateCollectionHidden(ctx context.Context, collectionID persist.DBID, hidden bool) error {
 	// Validate
-	if err := inputcheck.ValidateFields(api.validator, inputcheck.ValidationMap{
+	if err := validate.ValidateFields(api.validator, validate.ValidationMap{
 		"collectionID": {collectionID, "required"},
 	}); err != nil {
 		return err
